Workshop 2 Failed
"Class Diagram: View should Realize the IView interface (not generalization/specialization). Missing role names, and associations in the class diagram (Registry -> Member). Sequence Diagrams: In Add member there are calls in reg.addMember that are missing and wrong, i.e. the new Member message is sent from the reg:Registry object not the :Console. Also there is call to generateId and Member.setId that are missing. Retrieve member looks ok. Implementation: You still have hidden dependencies between the Controller and View class. E.g. if I change the function View.startPageHeader so that Show verbose list is the first option and show Verbose list is the second option I would need to change the corresponding switch in the Controller.readOptions operation. This is easy to spot as you have a hard coded constants in the View that displays the options e.g. the ""1."" in ""| 1. Add member(s) |"" this needs to match the case in the controller: e.g. ""case 1: return Options.ADD_MEMBER;"". This is extensively discussed in the suggested solution."